// server/src/modules/metadata/extractors/abs-metadata.types.ts
export interface AbsBookMetadata {
  title?: unknown;
  subtitle?: unknown;
  authors?: unknown;
  narrators?: unknown;
  series?: unknown;
  genres?: unknown;
  tags?: unknown;
  publishedYear?: unknown;
  publishedDate?: unknown;
  publisher?: unknown;
  description?: unknown;
  isbn?: unknown;
  asin?: unknown;
  language?: unknown;
  explicit?: unknown;
  abridged?: unknown;
  chapters?: unknown;
  metadata?: unknown;
}